2020/4/24 10:39

+++ Ruoyi/ruoyi-DS/src/main/java/com/ruoyi/DS/mapper/SkuproductMapper.java
@@ -17,7 +17,7 @@ public interface SkuproductMapper
      * @param id SKU单品ID
      * @return SKU单品
      */
-    public Skuproduct selectSkuproductById(Long id);
+    public Skuproduct selectSkuproductById(Integer id);
 
 
     /**
